refactor: move to domain, view package
