+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,8 @@
 package racingcar;
 
+import racingcar.ui.RacingController;
+import racingcar.view.InputView;
+
 public class Application {
 
     public static void main(String[] args) {
