+++ src/main/java/racingcar/domain/Cars.java
@@ -1,12 +1,20 @@
-package racingcar;
+package racingcar.domain;
+
+import racingcar.utils.StringUtils;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
 public class Cars {
+    private static final String CAR_NAME_REQUIRED_ERROR = "자동차 이름은 값이 존재해야 합니다.";
+
     private final List<Car> cars;
     List<Integer> positions = new ArrayList<>();
+
+    public Cars(String carNames) {
+        cars = createCar(carNames);
+    }
 
     public Cars(List<Car> cars) {
         this(cars, new ArrayList<>());
@@ -34,6 +42,20 @@ public class Cars {
 
     public List<Integer> getPositions() {
         return new ArrayList<>(positions);
+    }
+
+    private List<Car> createCar(String carNames) {
+        List<Car> tempCars = new ArrayList<>();
+
+        if (StringUtils.isBlank(carNames)) {
+            throw new IllegalArgumentException(CAR_NAME_REQUIRED_ERROR);
+        }
+        String[] names = carNames.split(",");
+
+        for (String name : names) {
+            tempCars.add(new Car(name));
+        }
+        return tempCars;
     }
 
     @Override
