+++ src/main/java/racingcar/domain/Number.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class Number {
     private final int postiveNumber;
