+++ src/main/java/racingcar/domain/Record.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.HashMap;
 import java.util.List;
