+++ src/test/java/racingcar/CarsTest.java
@@ -3,6 +3,8 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
 
 import java.util.Arrays;
