+++ src/test/java/racingcar/WinnersTest.java
@@ -2,6 +2,8 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.Winners;
 
 import java.util.Arrays;
 import java.util.List;
